Refactoring the modules
