--- integration-test-suite/accelerator-tests/gateway-integration/src/test/groovy/org/wso2/financial/services/accelerator/test/gateway/integration/accounts/util/AccountConstants.groovy
@@ -1,33 +1,0 @@
-/**
- * Copyright (c) 2024, WSO2 LLC. (https://www.wso2.com).
- *
- * WSO2 LLC. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.financial.services.accelerator.test.gateway.integration.accounts.util;
-
-/**
- * Common Constants of Account scenarios.
- */
-class AccountConstants {
-
-    static final String AISP_PATH = "open-banking/v3.1/aisp/"
-    static final String ACCOUNT_CONSENT_PATH = AISP_PATH + "account-access-consents"
-    static final String accountID = "30080012343456"
-    static final String ACCOUNTS_PATH = AISP_PATH + "accounts/" + accountID
-    static final String ACCOUNTS_BULK_PATH = AISP_PATH + "accounts"
-    static final String BALANCES_SINGLE_PATH = AISP_PATH + "accounts/" + accountID + "/balances"
-    static final String TRANSACTIONS_SINGLE_PATH = AISP_PATH + "accounts/" + accountID + "/transactions"
-}
