--- integration-test-suite/accelerator-tests/gateway-integration/src/test/groovy/org/wso2/financial/services/accelerator/test/gateway/integration/accounts/util/AccountPayloads.groovy
@@ -1,222 +1,0 @@
-/**
- * Copyright (c) 2024, WSO2 LLC. (https://www.wso2.com).
- *
- * WSO2 LLC. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.financial.services.accelerator.test.gateway.integration.accounts.util
-
-import groovy.json.JsonOutput
-import org.wso2.financial.services.accelerator.test.framework.constant.AcceleratorTestConstants
-
-import java.time.OffsetDateTime
-
-/**
- * Payloads for Account scenarios.
- */
-class AccountPayloads {
-
-    static String[] permissionsArray = [
-            AcceleratorTestConstants.READ_ACCOUNTS_BASIC,
-            AcceleratorTestConstants.READ_ACCOUNTS_DETAIL,
-            AcceleratorTestConstants.READ_BALANCES,
-            AcceleratorTestConstants.READ_TRANSACTIONS_DETAIL
-    ]
-
-    public static String initiationPayloadWithoutExpirationDate = """
-            {
-                "Data":{
-                "Permissions": ${JsonOutput.toJson(permissionsArray)},           
-                "TransactionFromDateTime":"${AcceleratorTestConstants.fromInstant}",
-                "TransactionToDateTime":"${AcceleratorTestConstants.toInstant}"
-            },
-                "Risk":{
-
-                }
-            }
-    """.stripIndent()
-
-    public static String initiationPayloadEmptyPayload = """""".stripIndent()
-    public static String initiationPayloadWithEmptyString = '""'
-    public static String initiationPayloadEmptyJsonPayload= """
-        {}
-    """.stripIndent()
-
-    public static String initiationPayloadWithInvalidExpirationDate = """
-        {
-            "Data":{
-            "Permissions": ${JsonOutput.toJson(permissionsArray)},
-            "ExpirationDateTime":"${AcceleratorTestConstants.fromInstant.toEpochSecond()}",
-            "TransactionFromDateTime":"${AcceleratorTestConstants.fromInstant}",
-            "TransactionToDateTime":"${AcceleratorTestConstants.toInstant}"
-        },
-            "Risk":{
-
-            }
-        }
-    """.stripIndent()
-
-    public static String initiationPayloadWithPastExpirationDate = """
-        {
-            "Data":{
-            "Permissions": ${JsonOutput.toJson(permissionsArray)},
-            "ExpirationDateTime":"${OffsetDateTime.now().minusDays(3)}",
-            "TransactionFromDateTime":"${AcceleratorTestConstants.fromInstant}",
-            "TransactionToDateTime":"${AcceleratorTestConstants.toInstant}"
-        },
-            "Risk":{
-
-            }
-        }
-    """.stripIndent()
-
-    public static String initiationPayloadWithoutTransactionToDate = """
-        {
-            "Data":{
-            "Permissions": ${JsonOutput.toJson(permissionsArray)},
-            "ExpirationDateTime":"${AcceleratorTestConstants.expirationInstant}",
-            "TransactionFromDateTime":"${AcceleratorTestConstants.fromInstant}",
-            "Risk":{
-
-            }
-        }
-    """.stripIndent()
-
-
-    public static String initiationPayloadWithoutTransactionFromDate = """
-        {
-            "Data":{
-            "Permissions": ${JsonOutput.toJson(permissionsArray)},
-            "ExpirationDateTime":"${AcceleratorTestConstants.expirationInstant}",
-            "TransactionToDateTime":"${AcceleratorTestConstants.toInstant}"
-        },
-            "Risk":{
-
-            }
-        }
-    """.stripIndent()
-
-    public static String initiationPayloadWithInvalidTransactionToDate = """
-        {
-            "Data":{
-            "Permissions": ${JsonOutput.toJson(permissionsArray)},
-            "ExpirationDateTime":"${AcceleratorTestConstants.expirationInstant}",
-            "TransactionFromDateTime":"${AcceleratorTestConstants.fromInstant}",
-            "TransactionToDateTime":"${AcceleratorTestConstants.toInstant.toEpochSecond()}"
-        },
-            "Risk":{
-
-            }
-        }
-    """.stripIndent()
-
-    public static String initiationPayloadWithInvalidTransactionFromDate = """
-        {
-            "Data":{
-            "Permissions": ${JsonOutput.toJson(permissionsArray)},
-            "ExpirationDateTime":"${AcceleratorTestConstants.expirationInstant}",
-            "TransactionFromDateTime":"${AcceleratorTestConstants.fromInstant.toEpochSecond()}",
-            "TransactionToDateTime":"${AcceleratorTestConstants.toInstant}"
-        },
-            "Risk":{
-
-            }
-        }
-    """.stripIndent()
-
-    public static String initiationPayloadWithPastTransactionToDateTime = """
-        {
-            "Data":{
-            "Permissions": ${JsonOutput.toJson(permissionsArray)},
-            "ExpirationDateTime":"${AcceleratorTestConstants.expirationInstant}",
-            "TransactionFromDateTime":"${AcceleratorTestConstants.fromInstant}",
-            "TransactionToDateTime":"${OffsetDateTime.now().minusDays(3)}"
-        },
-            "Risk":{
-
-            }
-        }
-    """.stripIndent()
-
-    public static String initiationPayloadWithSameTransactionFromToDates = """
-            {
-                "Data":{
-                "Permissions": ${JsonOutput.toJson(permissionsArray)},
-                "ExpirationDateTime":"${AcceleratorTestConstants.expirationInstant}",
-                "TransactionFromDateTime":"${AcceleratorTestConstants.fromInstant}",
-                "TransactionToDateTime":"${AcceleratorTestConstants.fromInstant}"
-            },
-                "Risk":{
-
-                }
-            }
-    """.stripIndent()
-
-    public static String initiationPayloadWithoutReadAccountBasic = """
-        {
-            "Data":{
-            "Permissions": ${JsonOutput.toJson([permissionsArray[1], permissionsArray[2], permissionsArray[3]])},
-            "ExpirationDateTime":"${AcceleratorTestConstants.expirationInstant}",
-            "TransactionFromDateTime":"${AcceleratorTestConstants.fromInstant}",
-            "TransactionToDateTime":"${AcceleratorTestConstants.toInstant}"
-        },
-            "Risk":{
-
-            }
-        }
-    """.stripIndent()
-
-    public static String initiationPayloadWithoutReadAccountDetails = """
-        {
-            "Data":{
-            "Permissions": ${JsonOutput.toJson([permissionsArray[0], permissionsArray[2], permissionsArray[3]])},
-            "ExpirationDateTime":"${AcceleratorTestConstants.expirationInstant}",
-            "TransactionFromDateTime":"${AcceleratorTestConstants.fromInstant}",
-            "TransactionToDateTime":"${AcceleratorTestConstants.toInstant}"
-        },
-            "Risk":{
-
-            }
-        }
-    """.stripIndent()
-
-    public static String initiationPayloadWithoutReadAccountBalance = """
-        {
-            "Data":{
-            "Permissions": ${JsonOutput.toJson([permissionsArray[0], permissionsArray[1], permissionsArray[3]])},
-            "ExpirationDateTime":"${AcceleratorTestConstants.expirationInstant}",
-            "TransactionFromDateTime":"${AcceleratorTestConstants.fromInstant}",
-            "TransactionToDateTime":"${AcceleratorTestConstants.toInstant}"
-        },
-            "Risk":{
-
-            }
-        }
-    """.stripIndent()
-
-    public static String initiationPayloadWithoutReadAccountTransactions = """
-        {
-            "Data":{
-            "Permissions": ${JsonOutput.toJson([permissionsArray[0], permissionsArray[1], permissionsArray[2]])},
-            "ExpirationDateTime":"${AcceleratorTestConstants.expirationInstant}",
-            "TransactionFromDateTime":"${AcceleratorTestConstants.fromInstant}",
-            "TransactionToDateTime":"${AcceleratorTestConstants.toInstant}"
-        },
-            "Risk":{
-
-            }
-        }
-    """.stripIndent()
-}
