+++ integration-test-suite/accelerator-tests/integration/test/accounts/src/test/resources/testng.xml
@@ -1,0 +1,41 @@
+<!--
+ ~ Copyright (c) 2024, WSO2 LLC. (https://www.wso2.com).
+ ~
+ ~ WSO2 LLC. licenses this file to you under the Apache License,
+ ~ Version 2.0 (the "License"); you may not use this file except
+ ~ in compliance with the License.
+ ~ You may obtain a copy of the License at
+ ~
+ ~     http://www.apache.org/licenses/LICENSE-2.0
+ ~
+ ~ Unless required by applicable law or agreed to in writing,
+ ~ software distributed under the License is distributed on an
+ ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ ~ KIND, either express or implied. See the License for the
+ ~ specific language governing permissions and limitations
+ ~ under the License.
+ -->
+
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
+<suite name="WSO2 Open Banking - Gateway Account Flow Test">
+    <test name="Accounts End to End Flow Tests" parallel="tests" enabled="true" group-by-instances="true">
+        <classes>
+            <class name="org.wso2.financial.services.accelerator.integration.test.accounts.EndToEndFlowTest.EndToEndAccountFlowTest" />
+        </classes>
+    </test>
+    <test name="Accounts Initiation Flow Tests" parallel="tests" enabled="true" group-by-instances="true">
+        <classes>
+            <class name="org.wso2.financial.services.accelerator.integration.test.accounts.Accounts_Initiation_Tests.AccountsInitiationRequestHeaderValidationTests" />
+            <class name="org.wso2.financial.services.accelerator.integration.test.accounts.Accounts_Initiation_Tests.AccountInitiationPayloadValidationTests" />
+            <class name="org.wso2.financial.services.accelerator.integration.test.accounts.Accounts_Initiation_Tests.GetConsentRequestHeaderValidationTests" />
+            <class name="org.wso2.financial.services.accelerator.integration.test.accounts.Accounts_Initiation_Tests.DeleteConsentRequestHeaderValidationTests" />
+            <class name="org.wso2.financial.services.accelerator.integration.test.accounts.Accounts_Initiation_Tests.DeleteConsentRequestValidationTests" />
+        </classes>
+    </test>
+    <test name="Accounts Retrieval Flow Tests" parallel="tests" enabled="true" group-by-instances="true">
+        <classes>
+            <class name="org.wso2.financial.services.accelerator.integration.test.accounts.Accounts_Retrieval_Tests.AccountsRetrievalRequestHeaderValidationTests" />
+            <class name="org.wso2.financial.services.accelerator.integration.test.accounts.Accounts_Retrieval_Tests.AccountsRetrievalRequestBodyValidationTests" />
+        </classes>
+    </test>
+</suite>
