+++ integration-test-suite/accelerator-tests/keymanager/test/consent-management/src/test/resources/testng.xml
@@ -1,0 +1,32 @@
+<!--
+ ~ Copyright (c) 2024, WSO2 LLC. (https://www.wso2.com).
+ ~
+ ~ WSO2 LLC. licenses this file to you under the Apache License,
+ ~ Version 2.0 (the "License"); you may not use this file except
+ ~ in compliance with the License.
+ ~ You may obtain a copy of the License at
+ ~
+ ~     http://www.apache.org/licenses/LICENSE-2.0
+ ~
+ ~ Unless required by applicable law or agreed to in writing,
+ ~ software distributed under the License is distributed on an
+ ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ ~ KIND, either express or implied. See the License for the
+ ~ specific language governing permissions and limitations
+ ~ under the License.
+ -->
+
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
+<suite name="WSO2 Open Banking - Consent Management Flow Test">
+    <test name="Consent Management Flow" parallel="tests" enabled="true" group-by-instances="true">
+        <classes>
+            <class name="org.wso2.financial.services.accelerator.keymanager.test.consent.management.EndToEndTest.EndToEndConsentManagementFlowTestFactory"/>
+            <class name="org.wso2.financial.services.accelerator.keymanager.test.consent.management.ConsentInitiationTest.ConsentInitiationTest"/>
+            <class name="org.wso2.financial.services.accelerator.keymanager.test.consent.management.ConsentRetrieval.ConsentRetrievalFlow"/>
+            <class name="org.wso2.financial.services.accelerator.keymanager.test.consent.management.ConsentDeleteTest.ConsentRevokeFlow"/>
+            <class name="org.wso2.financial.services.accelerator.keymanager.test.consent.management.ConsentValidateTest.ConsentValidationFlow"/>
+            <class name="org.wso2.financial.services.accelerator.keymanager.test.consent.management.ConsentValidateTest.ConsentManageEndpointValidationTest"/>
+            <class name="org.wso2.financial.services.accelerator.keymanager.test.consent.management.ConsentAuthorisationTest.AuthorizationFlowValidationTest"/>
+        </classes>
+    </test>
+</suite>
