+++ integration-test-suite/accelerator-tests/keymanager/test/token/src/test/groovy/grant_type_validation/ClientCredentialGrantAccessTokenTest.groovy
